feat: move ai service into feature module
